Revert, move to specs and publish the typescript vesion (#3769)
* Another try

* Do not transpile specs

* Another one

* Update snapshot
