+++ package.json
@@ -56,7 +56,7 @@
     "react-native-redash": "^12.0.3",
     "semver": "^5.5.0",
     "tinycolor2": "^1.4.2",
-    "uilib-native": "5.0.0-snapshot.7207",
+    "uilib-native": "5.0.0-snapshot.7212",
     "url-parse": "^1.2.0",
     "wix-react-native-text-size": "1.0.9"
   },
